FLORAL: Fit Log-ratio lasso regression for compositional covariates

Conduct log-ratio lasso regression for continuous, binary and survival outcomes.

Usage

FLORAL(
  x,
  y,
  family = "gaussian",
  longitudinal = FALSE,
  id = NULL,
  tobs = NULL,
  failcode = NULL,
  length.lambda = 100,
  lambda.min.ratio = NULL,
  mu = 1,
  ncv = 5,
  intercept = FALSE,
  foldid = NULL,
  step2 = TRUE,
  progress = TRUE,
  plot = TRUE
)

Arguments

x

Count data matrix, where rows specify subjects and columns specify features. If `x` contains longitudinal data, the rows must be sorted in the same order of the subject IDs used in `y`.

y

Outcome. For a continuous or binary outcome, `y` is a vector. For survival outcome, `y` is a `Surv` object.

family

Available options are `gaussian`, `binomial`, `cox`, `finegray`. `finegray` is still under development. Please use with caution.

longitudinal

TRUE or FALSE, indicating whether longitudinal data matrix is specified for input `x`. (Still under development. Please use with caution)

id

If `longitudinal` is TRUE, `id` specifies subject IDs corresponding to the rows of input `x`.

tobs

If `longitudinal` is TRUE, `tobs` specifies time points corresponding to the rows of input `x`.

failcode

If `family = finegray`, `failcode` specifies the failure type of interest. This must be a positive integer.

length.lambda

Number of penalty parameters used in the path

lambda.min.ratio

Ratio between the minimum and maximum choice of lambda. Default is `NULL`, where the ratio is chosen as 1e-2 if n < p and 1e-4 otherwise.

mu

Value of penalty for the augmented Lagrangian

ncv

Number of cross-validation runs. Use `NULL` if cross-validation is not wanted.

intercept

TRUE or FALSE, indicating whether an intercept should be estimated.

foldid

A vector of fold indicator. Default is `NULL`.

step2

TRUE or FALSE, indicating whether a second-stage feature selection for specific ratios should be performed for the features selected by the main lasso algorithm. Will only be performed if cross validation is enabled.

progress

TRUE or FALSE, indicating whether printing progress bar as the algorithm runs.

plot

TRUE or FALSE, indicating whether returning plots of model fitting.

Value

A list with path-specific estimates (beta), path (lambda), and many others.

Author

Teng Fei. Email: feit1@mskcc.org

Examples

set.seed(23420)

# Continuous outcome
dat <- simu(n=50,p=30,model="linear")
fit <- FLORAL(dat$xcount,dat$y,family="gaussian",progress=FALSE,step2=TRUE)

# Binary outcome
dat <- simu(n=50,p=30,model="binomial")
fit <- FLORAL(dat$xcount,dat$y,family="binomial",progress=FALSE,step2=TRUE)

# Survival outcome
dat <- simu(n=50,p=30,model="cox")
fit <- FLORAL(dat$xcount,survival::Surv(dat$t,dat$d),family="cox",progress=FALSE,step2=TRUE)

# Competing risks outcome
dat <- simu(n=50,p=30,model="finegray")
fit <- FLORAL(dat$xcount,survival::Surv(dat$t,dat$d,type="mstate"),failcode=1,family="finegray",progress=FALSE,step2=FALSE)